#620: Refactor events plugin
